refactored tests
- extracted common test helpers
- separated tests related to range-init
